[Projects] Separation of core Project resource from project orgs, project roles, project markings
